textfussion/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
blessed==1.20.0
brotlipy==0.7.0
cachetools==5.3.0
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
datasets==2.11.0
decorator==5.1.1
deepspeed==0.8.3
diffusers==0.15.0.dev0
dill==0.3.6
einops==0.6.1
executing==1.2.0
filelock==3.10.7
flit_core==3.8.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
future==0.18.3
google-auth==2.17.1
google-auth-oauthlib==1.0.0
gpustat==1.1
grpcio==1.53.0
hjson==3.1.0
huggingface-hub==0.13.3
idna==3.4
imageio==2.27.0
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==6.1.0
ipdb==0.13.13
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonlines==3.1.0
kiwisolver==1.4.4
lazy_loader==0.2
lit==16.0.0
lmdb==1.4.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==2.0.0
mmdet==3.0.0
mmengine==0.7.3
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.0
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.1.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.0.1
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyre-extensions==0.0.29
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.5
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
setuptools==65.6.3
shapely==2.0.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminaltables==3.1.10
tifffile==2023.3.21
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
triton==2.0.0
typing_extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wheel==0.38.4
xformers==0.0.18+bde8248.d20230406
xmltodict==0.13.0
xxhash==3.2.0
yapf==0.33.0
yarl==1.8.2
zipp==3.15.0
